+++ plugins/action/cf_service_instance_info.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 #
 # SPDX-License-Identifier: GPL-3.0-only
-# SPDX-FileCopyrightText: 2023 Red Hat, Project Atmosphere
+# SPDX-FileCopyrightText: 2024 Red Hat, Project Atmosphere
 #
-# Copyright 2023 Red Hat, Project Atmosphere
+# Copyright 2024 Red Hat, Project Atmosphere
 #
 # This program is free software: you can redistribute it and/or modify it under the terms of the GNU
 # General Public License as published by the Free Software Foundation, version 3 of the License.
@@ -27,85 +27,21 @@ from ansible.errors import AnsibleActionFail
 
 __metaclass__ = type
 import json
-import subprocess  # nosec B404
 from ansible.plugins.action import ActionBase
 
-
-class CFClient:
-    def __init__(
-        self,
-        api_endpoint: str = "",
-        username: str = "",
-        password: str = "",  # nosec B107
-        cf_home: str = "",
-    ):
-        """Initializes the CFClient object with optional keyword arguments."""
-        self.api_endpoint = api_endpoint
-        self.username = username
-        self.password = password
-        self.cf_home = cf_home
-        self.env = {"CF_HOME": cf_home}
-
-    def authenticate(self):
-        rc = subprocess.run(
-            [
-                f'cf login -a "{self.api_endpoint}" -u "{self.username}" -p "{self.password}"',
-            ],
-            shell=True,  # nosec B602
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=False,
-            env=self.env,
-        )
-        if rc.returncode:
-            return {
-                "failed": True,
-                "msg": "Error while authorizing cf cli",
-            }
-        else:
-            return {"failed": False}
-
-    def run_method(
-        self,
-        path: str,
-        data: dict,
-        method: str = "GET",
-        headers: dict[str, str] = None,
-    ):
-        rc = self.authenticate()
-        if rc.get("failed"):
-            return rc
-
-        if headers is None:
-            headers = {"Content-Type": "application/x-www-form-urlencoded"}
-        # headers_command = [f'-H "{k}:{v}"' for k, v in headers]
-        data_string = json.dumps(data)
-        command_run = subprocess.run(
-            [
-                f"""cf curl "{path}" -X {method} -d '{data_string}'""",
-            ],
-            # + headers_command,
-            shell=True,  # nosec B602
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=False,
-            env=self.env,
-        )
-
-        return (
-            command_run.returncode,
-            command_run.stdout.decode("utf-8"),
-            command_run.stderr.decode("utf-8"),
-        )
+from ansible_collections.sap.sap_operations.plugins.action.cf_marketplace_info import (
+    CFClient,
+)
 
 
-def run_action(client: CFClient) -> dict:
+def run_action(client: CFClient, name: str = "", guid: str = "") -> dict:
+
     rc = client.authenticate()
     if rc.get("failed"):
         return rc
 
     rc, stdout, stderr = client.run_method(
-        path="/v3/service_offerings",
+        path="/v3/service_instances",
         method="GET",
         headers={},
         data={},
@@ -114,24 +50,42 @@ def run_action(client: CFClient) -> dict:
     if rc or stderr:
         return {"failed": True, "msg": stdout}
 
-    result = json.loads(stdout)
+    service_instances = json.loads(stdout)
 
-    if "errors" in result:
+    if "errors" in service_instances:
         return {
             "failed": True,
-            "msg": "Error retrieving service offerings",
-            "errors": result["errors"],
+            "msg": "Error retrieving service instances",
+            "errors": service_instances["errors"],
         }
+    for service_instance in service_instances["resources"]:
+        if name and service_instance["name"] == name:
+            return dict(
+                failed=False,
+                changed=False,
+                cf_service_instance_info=service_instance,
+            )
+        if guid and service_instance["guid"] == guid:
+            return dict(
+                failed=False,
+                changed=False,
+                cf_service_instance_info=service_instance,
+            )
 
-    return dict(failed=False, changed=False, cf_marketplace_info=result["resources"])
+    return dict(
+        failed=True,
+        msg="Service instance not found",
+    )
 
 
-class CFMarketplaceInfoActionModule(ActionBase):
-    _VALID_ARGS = frozenset(["username", "password", "api_endpoint"])
+class CFServiceInstanceInfoModule(ActionBase):
+    _VALID_ARGS = frozenset(["username", "password", "api_endpoint", "name", "guid"])
     argument_spec = dict(
         username=dict(type="str", required=False),
         password=dict(type="str", required=False),
         api_endpoint=dict(type="str", required=False),
+        name=dict(type="str", required=False),
+        guid=dict(type="str", required=False),
     )
     required_together = [
         [
@@ -140,13 +94,14 @@ class CFMarketplaceInfoActionModule(ActionBase):
             "api_endpoint",
         ],
     ]
-    mutually_exclusive = []
+    mutually_exclusive = [["name", "guid"]]
     required_one_of = [
         [
             "username",
             "password",
             "api_endpoint",
-        ]
+        ],
+        ["name", "guid"],
     ]
 
     def __init__(
@@ -174,6 +129,8 @@ class CFMarketplaceInfoActionModule(ActionBase):
         api_endpoint = self._task.args.get("api_endpoint")
         username = self._task.args.get("username")
         password = self._task.args.get("password")
+        name = self._task.args.get("name", "")
+        guid = self._task.args.get("guid", "")
 
         with tempfile.TemporaryDirectory() as cf_home:
             client = CFClient(
@@ -182,8 +139,8 @@ class CFMarketplaceInfoActionModule(ActionBase):
                 password=password,
                 cf_home=cf_home,
             )
-            return run_action(client)
+            return run_action(client, name=name, guid=guid)
 
 
-class ActionModule(CFMarketplaceInfoActionModule):
+class ActionModule(CFServiceInstanceInfoModule):
     pass