+++ plugins/doc_fragments/cloud_foundry.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: GPL-3.0-only
-# SPDX-FileCopyrightText: 2023 Red Hat, Project Atmosphere
+# SPDX-FileCopyrightText: 2024 Red Hat, Project Atmosphere
 #
-# Copyright 2023 Red Hat, Project Atmosphere
+# Copyright 2024 Red Hat, Project Atmosphere
 #
 # This program is free software: you can redistribute it and/or modify it under the terms of the GNU
 # General Public License as published by the Free Software Foundation, version 3 of the License.
@@ -19,13 +19,30 @@
 # You should have received a copy of the GNU General Public License along with this program.
 # If not, see <https://www.gnu.org/licenses/>.
 
+
+class ModuleDocFragment(object):
+    DOCUMENTATION = r"""
 ---
-requires_ansible: ">=2.14.0"
-action_groups:
-  cf:
-    - sap.sap_operations.cf_marketplace_info
-    - sap.sap_operations.cf_service_instance
-    - sap.sap_operations.cf_service_instance_info
-    - sap.sap_operations.cf_service_instances_info
-    - sap.sap_operations.cf_plans_info
-    - sap.sap_operations.cf_spaces_info
+requirements:
+  - "python >= 3.6"
+  - cf CLI should be installed and available in PATH
+author:
+  - Kirill Satarin (@kksat)
+options:
+  username:
+    description:
+      - The username for authentication with the Cloud Foundry API.
+      - This is SAP BTP user email address
+    type: str
+    required: false
+  password:
+    description:
+      - The password for authentication with the Cloud Foundry API.
+    type: str
+    required: false
+  api_endpoint:
+    description:
+      - The endpoint URL of the Cloud Foundry API.
+    type: str
+    required: false
+"""